pyproject.toml
[tool.poetry]
name = "finodays"
version = "0.1.0"
description = ""
authors = ["kirill <shkolenko02-03-2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.0"}
loguru = "^0.7.0"
alembic = "^1.11.3"
psycopg2-binary = "^2.9.7"
uuid = "^1.30"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
